Move object mapper out to a helper
